feat(Gyoungmin-Min):[BOJ, 1874] 스택수열 https://www.acmicpc.net/problem/1874
